requirements.txt
python-pptx>=1.0.0
google-genai>=1.0.0
click>=8.1.7
python-dotenv>=1.0.1
lxml>=5.2.0
Flask>=2.3.0
Werkzeug>=2.3.0
gunicorn>=21.0.0
